New Method for Grabbing Title
Changed it so that instead of doing the complicated thing of referencing "Music:" - concerning when Gilva Smash Bros. formatting is all weird - it now references the first ```. There's a better way to find the first " ``` " than limiting the amount of values you can index, but I'll put that in later.