Modified Indentation, Safety Guards
Created a safety guard in the case that "by me" is found in the line of a rip. It will still print out the name (in case a ripper's name begins with me, e.g. Metalik), along with the Discord tag. Also, made the indentation less unorthodox; repl.it's fault.